Make benchmarks pass based on 'All' or 'Any' mode
Every benchmark in an audit spec has a field 'Mode' which can have
the value 'All' (default) or 'Any'. If the value is 'All' then all
audit commands given in the 'run' steps must pass. If the value is
'Any', then the benchmark as a whole passes when any one of them
passes.